CmdAdd: check that VF has valid GUID
Currently ib-sriov-cni configures GUID according to the value provided
in the CNI's network configuration. This value is either set by
ib-kubernetes or by other means (e.g, by manually specifying it in the
CNI configuration). When GUID is not specified, the existing GUID that
is currently set on the VF is used. Verify VF's GUID is valid in such
case.